Move types to common file, remove unnecessary documentation
- move types from TextInput, TextInputFlat, TextInputOutlined to common file,
- remove documentation from TextInputFlat and TextInputOutlined
