+++ src/components/TextInput/TextInput.js
@@ -7,25 +7,10 @@ import { polyfill } from 'react-lifecycles-compat';
 import TextInputOutlined from './TextInputOutlined';
 import TextInputFlat from './TextInputFlat';
 import { withTheme } from '../../core/theming';
-import type { Theme } from '../../types';
+import type { Props } from './types';
 
-type RenderProps = {
-  ref: any => void,
-  onChangeText: string => void,
-  placeholder: ?string,
-  placeholderTextColor: string,
-  editable?: boolean,
-  selectionColor: string,
-  onFocus: () => mixed,
-  onBlur: () => mixed,
-  underlineColorAndroid: string,
-  style: any,
-  multiline?: boolean,
-  numberOfLines?: number,
-  value?: string,
-};
-
-type Props = React.ElementConfig<typeof NativeTextInput> & {|
+type TextInputProps = {|
+  ...Props,
   /**
    * Mode of the TextInput.
    * - `flat` - flat input with an underline.
@@ -35,78 +20,6 @@ type Props = React.ElementConfig<typeof NativeTextInput> & {|
    * This component render TextInputOutlined or TextInputFlat based on that props
    */
   mode?: 'flat' | 'outlined',
-  /**
-   * If true, user won't be able to interact with the component.
-   */
-  disabled?: boolean,
-  /**
-   * The text to use for the floating label.
-   */
-  label?: string,
-  /**
-   * Placeholder for the input.
-   */
-  placeholder?: string,
-  /**
-   * Whether to style the TextInput with error style.
-   */
-  error?: boolean,
-  /**
-   * Callback that is called when the text input's text changes. Changed text is passed as an argument to the callback handler.
-   */
-  onChangeText?: Function,
-  /**
-   * Selection color of the input
-   */
-  selectionColor?: string,
-  /**
-   * Underline color of the input.
-   */
-  underlineColor?: string,
-  /**
-   * Whether the input can have multiple lines.
-   */
-  multiline?: boolean,
-  /**
-   * The number of lines to show in the input (Android only).
-   */
-  numberOfLines?: number,
-  /**
-   * Callback that is called when the text input is focused.
-   */
-  onFocus?: () => mixed,
-  /**
-   * Callback that is called when the text input is blurred.
-   */
-  onBlur?: () => mixed,
-  /**
-   *
-   * Callback to render a custom input component such as `react-native-text-input-mask`
-   * instead of the default `TextInput` component from `react-native`.
-   *
-   * Example:
-   * ```js
-   * <TextInput
-   *   label="Phone number"
-   *   render={props =>
-   *     <TextInputMask
-   *       {...props}
-   *       mask="+[00] [000] [000] [000]"
-   *     />
-   *   }
-   * />
-   * ```
-   */
-  render: (props: RenderProps) => React.Node,
-  /**
-   * Value of the text input.
-   */
-  value?: string,
-  style?: any,
-  /**
-   * @optional
-   */
-  theme: Theme,
 |};
 
 type State = {};
@@ -158,7 +71,7 @@ type State = {};
  * @extends TextInput props https://facebook.github.io/react-native/docs/textinput.html#props
  */
 
-class TextInput extends React.Component<Props, State> {
+class TextInput extends React.Component<TextInputProps, State> {
   static defaultProps = {
     mode: 'flat',
     disabled: false,
