+++ src/components/TextInput/TextInputFlat.js
@@ -11,7 +11,7 @@ import {
 import { polyfill } from 'react-lifecycles-compat';
 import color from 'color';
 import Text from '../Typography/Text';
-import type { Theme } from '../../types';
+import type { Props, State } from './types';
 
 const AnimatedText = Animated.createAnimatedComponent(Text);
 
@@ -23,156 +23,6 @@ const FOCUS_ANIMATION_DURATION = 150;
 const BLUR_ANIMATION_DURATION = 180;
 const LABEL_PADDING_HORIZONTAL = 12;
 const RANDOM_VALUE_TO_CENTER_LABEL = 4; // Don't know why 4, but it works
-
-type RenderProps = {
-  ref: any => void,
-  onChangeText: string => void,
-  placeholder: ?string,
-  placeholderTextColor: string,
-  editable?: boolean,
-  selectionColor: string,
-  onFocus: () => mixed,
-  onBlur: () => mixed,
-  underlineColorAndroid: string,
-  style: any,
-  multiline?: boolean,
-  numberOfLines?: number,
-  value?: string,
-};
-
-type Props = React.ElementConfig<typeof NativeTextInput> & {|
-  /**
-   * If true, user won't be able to interact with the component.
-   */
-  disabled?: boolean,
-  /**
-   * The text to use for the floating label.
-   */
-  label?: string,
-  /**
-   * Placeholder for the input.
-   */
-  placeholder?: string,
-  /**
-   * Whether to style the TextInput with error style.
-   */
-  error?: boolean,
-  /**
-   * Callback that is called when the text input's text changes. Changed text is passed as an argument to the callback handler.
-   */
-  onChangeText?: Function,
-  /**
-   * Selection color of the input
-   */
-  selectionColor?: string,
-  /**
-   * Underline color of the input.
-   */
-  underlineColor?: string,
-  /**
-   * Whether the input can have multiple lines.
-   */
-  multiline?: boolean,
-  /**
-   * The number of lines to show in the input (Android only).
-   */
-  numberOfLines?: number,
-  /**
-   * Callback that is called when the text input is focused.
-   */
-  onFocus?: () => mixed,
-  /**
-   * Callback that is called when the text input is blurred.
-   */
-  onBlur?: () => mixed,
-  /**
-   *
-   * Callback to render a custom input component such as `react-native-text-input-mask`
-   * instead of the default `TextInput` component from `react-native`.
-   *
-   * Example:
-   * ```js
-   * <TextInput
-   *   label="Phone number"
-   *   render={props =>
-   *     <TextInputMask
-   *       {...props}
-   *       mask="+[00] [000] [000] [000]"
-   *     />
-   *   }
-   * />
-   * ```
-   */
-  render: (props: RenderProps) => React.Node,
-  /**
-   * Value of the text input.
-   */
-  value?: string,
-  style?: any,
-  /**
-   * @optional
-   */
-  theme: Theme,
-|};
-
-type State = {
-  labeled: Animated.Value,
-  error: Animated.Value,
-  focused: boolean,
-  placeholder: ?string,
-  value: ?string,
-  labelLayout: {
-    measured: boolean,
-    width: number,
-  },
-};
-
-/**
- * A component to allow users to input text.
- *
- * <div class="screenshots">
- *   <figure>
- *     <img class="medium" src="screenshots/textinput-flat.focused.png" />
- *     <figcaption>Flat (focused)</figcaption>
- *   </figure>
- *   <figure>
- *     <img class="medium" src="screenshots/textinput-flat.disabled.png" />
- *     <figcaption>Flat (disabled)</figcaption>
- *   </figure>
- *   <figure>
- *     <img class="medium" src="screenshots/textinput-outlined.focused.png" />
- *     <figcaption>Outlined (focused)</figcaption>
- *   </figure>
- *   <figure>
- *     <img class="medium" src="screenshots/textinput-outlined.disabled.png" />
- *     <figcaption>Outlined (disabled)</figcaption>
- *   </figure>
- * </div>
- *
- * ## Usage
- * ```js
- * import * as React from 'react';
- * import { TextInput } from 'react-native-paper';
- *
- * export default class MyComponent extends React.Component {
- *   state = {
- *     text: ''
- *   };
- *
- *   render(){
- *     return (
- *       <TextInput
- *         label='Email'
- *         value={this.state.text}
- *         onChangeText={text => this.setState({ text })}
- *       />
- *     );
- *   }
- * }
- * ```
- *
- * @extends TextInput props https://facebook.github.io/react-native/docs/textinput.html#props
- */
 
 class TextInputFlat extends React.Component<Props, State> {
   static defaultProps = {
